[Decomp][Player] Cohesion split: extract durability methods to PlayerDurability.cpp
