+++ src/game/Object/Player.cpp
@@ -4646,325 +4646,12 @@ Corpse* Player::GetCorpse() const
     return sObjectAccessor.GetCorpseForPlayerGUID(GetObjectGuid());
 }
 
-/**
- * @brief Applies percentage-based durability loss to equipped items and optionally inventory items.
- *
- * @param percent The fraction of maximum durability to remove.
- * @param inventory True to include inventory and bag contents.
- */
-void Player::DurabilityLossAll(double percent, bool inventory)
-{
-    for (int i = EQUIPMENT_SLOT_START; i < EQUIPMENT_SLOT_END; ++i)
-    {
-        if (Item* pItem = GetItemByPos(INVENTORY_SLOT_BAG_0, i))
-        {
-            DurabilityLoss(pItem, percent);
-        }
-    }
 
-    if (inventory)
-    {
-        // bags not have durability
-        // for(int i = INVENTORY_SLOT_BAG_START; i < INVENTORY_SLOT_BAG_END; ++i)
 
-        for (int i = INVENTORY_SLOT_ITEM_START; i < INVENTORY_SLOT_ITEM_END; ++i)
-        {
-            if (Item* pItem = GetItemByPos(INVENTORY_SLOT_BAG_0, i))
-            {
-                DurabilityLoss(pItem, percent);
-            }
-        }
 
-        // keys not have durability
-        // for(int i = KEYRING_SLOT_START; i < KEYRING_SLOT_END; ++i)
 
-        for (int i = INVENTORY_SLOT_BAG_START; i < INVENTORY_SLOT_BAG_END; ++i)
-        {
-            if (Bag* pBag = (Bag*)GetItemByPos(INVENTORY_SLOT_BAG_0, i))
-                for (uint32 j = 0; j < pBag->GetBagSize(); ++j)
-                {
-                    if (Item* pItem = GetItemByPos(i, j))
-                    {
-                        DurabilityLoss(pItem, percent);
-                    }
-                }
-        }
-    }
-}
 
-/**
- * @brief Applies percentage-based durability loss to a single item.
- *
- * @param item The item to damage.
- * @param percent The fraction of maximum durability to remove.
- */
-void Player::DurabilityLoss(Item* item, double percent)
-{
-    if (!item)
-    {
-        return;
-    }
 
-    uint32 pMaxDurability =  item ->GetUInt32Value(ITEM_FIELD_MAXDURABILITY);
-
-    if (!pMaxDurability)
-    {
-        return;
-    }
-
-    uint32 pDurabilityLoss = uint32(pMaxDurability * percent);
-
-    if (pDurabilityLoss < 1)
-    {
-        pDurabilityLoss = 1;
-    }
-
-    DurabilityPointsLoss(item, pDurabilityLoss);
-}
-
-/**
- * @brief Applies flat durability loss to equipped items and optionally inventory items.
- *
- * @param points The number of durability points to remove.
- * @param inventory True to include inventory and bag contents.
- */
-void Player::DurabilityPointsLossAll(int32 points, bool inventory)
-{
-    for (int i = EQUIPMENT_SLOT_START; i < EQUIPMENT_SLOT_END; ++i)
-    {
-        if (Item* pItem = GetItemByPos(INVENTORY_SLOT_BAG_0, i))
-        {
-            DurabilityPointsLoss(pItem, points);
-        }
-    }
-    if (inventory)
-    {
-        // bags not have durability
-        // for(int i = INVENTORY_SLOT_BAG_START; i < INVENTORY_SLOT_BAG_END; ++i)
-
-        for (int i = INVENTORY_SLOT_ITEM_START; i < INVENTORY_SLOT_ITEM_END; ++i)
-        {
-            if (Item* pItem = GetItemByPos(INVENTORY_SLOT_BAG_0, i))
-            {
-                DurabilityPointsLoss(pItem, points);
-            }
-        }
-
-        // keys not have durability
-        // for(int i = KEYRING_SLOT_START; i < KEYRING_SLOT_END; ++i)
-
-        for (int i = INVENTORY_SLOT_BAG_START; i < INVENTORY_SLOT_BAG_END; ++i)
-        {
-            if (Bag* pBag = (Bag*)GetItemByPos(INVENTORY_SLOT_BAG_0, i))
-                for (uint32 j = 0; j < pBag->GetBagSize(); ++j)
-                {
-                    if (Item* pItem = GetItemByPos(i, j))
-                    {
-                        DurabilityPointsLoss(pItem, points);
-                    }
-                }
-        }
-    }
-}
-
-/**
- * @brief Applies flat durability loss to a single item.
- *
- * @param item The item to damage.
- * @param points The number of durability points to remove.
- */
-void Player::DurabilityPointsLoss(Item* item, int32 points)
-{
-    int32 pMaxDurability = item->GetUInt32Value(ITEM_FIELD_MAXDURABILITY);
-    int32 pOldDurability = item->GetUInt32Value(ITEM_FIELD_DURABILITY);
-    int32 pNewDurability = pOldDurability - points;
-
-    if (pNewDurability < 0)
-    {
-        pNewDurability = 0;
-    }
-    else if (pNewDurability > pMaxDurability)
-    {
-        pNewDurability = pMaxDurability;
-    }
-
-    if (pOldDurability != pNewDurability)
-    {
-        // modify item stats _before_ Durability set to 0 to pass _ApplyItemMods internal check
-        if (pNewDurability == 0 && pOldDurability > 0 && item->IsEquipped())
-        {
-            _ApplyItemMods(item, item->GetSlot(), false);
-        }
-
-        item->SetUInt32Value(ITEM_FIELD_DURABILITY, pNewDurability);
-
-        // modify item stats _after_ restore durability to pass _ApplyItemMods internal check
-        if (pNewDurability > 0 && pOldDurability == 0 && item->IsEquipped())
-        {
-            _ApplyItemMods(item, item->GetSlot(), true);
-        }
-
-        item->SetState(ITEM_CHANGED, this);
-    }
-}
-
-/**
- * @brief Removes one durability point from an equipped item slot.
- *
- * @param slot The equipment slot to damage.
- */
-void Player::DurabilityPointLossForEquipSlot(EquipmentSlots slot)
-{
-    if (Item* pItem = GetItemByPos(INVENTORY_SLOT_BAG_0, slot))
-    {
-        DurabilityPointsLoss(pItem, 1);
-    }
-}
-
-/**
- * @brief Repairs all eligible equipped and carried items.
- *
- * @param cost True to charge the player for repairs.
- * @param discountMod The vendor discount multiplier to apply.
- * @return The total repair cost.
- */
-uint32 Player::DurabilityRepairAll(bool cost, float discountMod, bool guildBank)
-{
-    uint32 TotalCost = 0;
-    // equipped, backpack, bags itself
-    for (int i = EQUIPMENT_SLOT_START; i < INVENTORY_SLOT_ITEM_END; ++i)
-    {
-        TotalCost += DurabilityRepair(((INVENTORY_SLOT_BAG_0 << 8) | i), cost, discountMod, guildBank);
-    }
-
-    // bank, buyback and keys not repaired
-
-    // items in inventory bags
-    for (int j = INVENTORY_SLOT_BAG_START; j < INVENTORY_SLOT_BAG_END; ++j)
-    {
-        for (int i = 0; i < MAX_BAG_SIZE; ++i)
-        {
-            TotalCost += DurabilityRepair(((j << 8) | i), cost, discountMod, guildBank);
-        }
-    }
-    return TotalCost;
-}
-
-/**
- * @brief Repairs a single item to full durability.
- *
- * @param pos The packed inventory position of the item.
- * @param cost True to charge the player for the repair.
- * @param discountMod The vendor discount multiplier to apply.
- * @return The repair cost for the item.
- */
-uint32 Player::DurabilityRepair(uint16 pos, bool cost, float discountMod, bool guildBank)
-{
-    Item* item = GetItemByPos(pos);
-
-    uint32 TotalCost = 0;
-    if (!item)
-    {
-        return TotalCost;
-    }
-
-    uint32 maxDurability = item->GetUInt32Value(ITEM_FIELD_MAXDURABILITY);
-    if (!maxDurability)
-    {
-        return TotalCost;
-    }
-
-    uint32 curDurability = item->GetUInt32Value(ITEM_FIELD_DURABILITY);
-
-    if (cost)
-    {
-        uint32 LostDurability = maxDurability - curDurability;
-        if (LostDurability > 0)
-        {
-            ItemPrototype const* ditemProto = item->GetProto();
-
-            DurabilityCostsEntry const* dcost = sDurabilityCostsStore.LookupEntry(ditemProto->ItemLevel);
-            if (!dcost)
-            {
-                sLog.outError("RepairDurability: Wrong item lvl %u", ditemProto->ItemLevel);
-                return TotalCost;
-            }
-
-            uint32 dQualitymodEntryId = (ditemProto->Quality + 1) * 2;
-            DurabilityQualityEntry const* dQualitymodEntry = sDurabilityQualityStore.LookupEntry(dQualitymodEntryId);
-            if (!dQualitymodEntry)
-            {
-                sLog.outError("RepairDurability: Wrong dQualityModEntry %u", dQualitymodEntryId);
-                return TotalCost;
-            }
-
-            uint32 dmultiplier = dcost->multiplier[ItemSubClassToDurabilityMultiplierId(ditemProto->Class, ditemProto->SubClass)];
-            uint32 costs = uint32(LostDurability * dmultiplier * double(dQualitymodEntry->quality_mod));
-
-            costs = uint32(costs * discountMod);
-
-            if (costs == 0)                                 // fix for ITEM_QUALITY_ARTIFACT
-            {
-                costs = 1;
-            }
-
-            if (guildBank)
-            {
-                if (GetGuildId() == 0)
-                {
-                    DEBUG_LOG("You are not member of a guild");
-                    return TotalCost;
-                }
-
-                Guild* pGuild = sGuildMgr.GetGuildById(GetGuildId());
-                if (!pGuild)
-                {
-                    return TotalCost;
-                }
-
-                if (!pGuild->HasRankRight(GetRank(), GR_RIGHT_WITHDRAW_REPAIR))
-                {
-                    DEBUG_LOG("You do not have rights to withdraw for repairs");
-                    return TotalCost;
-                }
-
-                if (pGuild->GetMemberMoneyWithdrawRem(GetGUIDLow()) < costs)
-                {
-                    DEBUG_LOG("You do not have enough money withdraw amount remaining");
-                    return TotalCost;
-                }
-
-                if (pGuild->GetGuildBankMoney() < costs)
-                {
-                    DEBUG_LOG("There is not enough money in bank");
-                    return TotalCost;
-                }
-
-                pGuild->MemberMoneyWithdraw(costs, GetGUIDLow());
-                TotalCost = costs;
-            }
-            else if (GetMoney() < costs)
-            {
-                DEBUG_LOG("You do not have enough money");
-                return TotalCost;
-            }
-            else
-            {
-                ModifyMoney(-int32(costs));
-            }
-        }
-    }
-
-    item->SetUInt32Value(ITEM_FIELD_DURABILITY, maxDurability);
-    item->SetState(ITEM_CHANGED, this);
-
-    // reapply mods for total broken and repaired item if equipped
-    if (IsEquipmentPos(pos) && !curDurability)
-    {
-        _ApplyItemMods(item, pos & 255, true);
-    }
-    return TotalCost;
-}
 
 /**
  * @brief Moves the player to the nearest valid graveyard.
